Modified with latest wpeframework.
